+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 Only major changes and additions will be noted here.
+- **2018-05-10** Added the original Google stock icons to the automated preview generator.
 - **2018-05-09** Added a section for [rejected sugestions](https://petershaggynoble.github.io/MDI-Sandbox/?section=rejected), for the benefit of those who might have use for them.
 - **2018-05-09:** Added an automated preview generator - see [README.md](README.md#preview-generator) for usage instructions.
 - **2018-05-08:** "My Favourites" renamed as "My Library".